Split up platform crate.
