+++ crates/core/dep-graph/src/dep_builder.rs
@@ -97,9 +97,8 @@ impl<'ws> DepGraphBuilder<'ws> {
         // If project is NOT in the package manager workspace, then we should
         // install dependencies in the project, not the workspace root.
         if let Some(platform) = self.platforms.get(project.language) {
-            if !platform.is_project_in_package_manager_workspace(&project.id, &project.root)? {
-                installs_in_project = true;
-            }
+            installs_in_project =
+                !platform.is_project_in_dependency_workspace(&project.id, &project.root)?;
         }
 
         // When installing dependencies in the project, we will use the
