+++ crates/core/platform/src/platform.rs
@@ -1,0 +1,68 @@
+use crate::runtime::Runtime;
+use moon_config::{
+    DependencyConfig, PlatformType, ProjectConfig, ProjectLanguage, ProjectsAliasesMap,
+    ProjectsSourcesMap, TasksConfigsMap,
+};
+use moon_error::MoonError;
+use std::collections::BTreeMap;
+use std::fmt::Debug;
+use std::path::Path;
+
+pub trait Platform: Debug + Send + Sync {
+    /// Detect the language of project at the provided path.
+    fn detect_project_language(&self, project_root: &Path) -> Option<ProjectLanguage> {
+        None
+    }
+
+    /// Return the type of this platform.
+    fn get_type(&self) -> PlatformType;
+
+    /// Return a runtime with an appropriate version based on the provided configs.
+    fn get_runtime_from_config(&self, project_config: Option<&ProjectConfig>) -> Option<Runtime>;
+
+    /// Determine if the provided project is within the platform's dependency manager
+    /// workspace (not to be confused with moon's workspace).
+    fn is_project_in_dependency_workspace(
+        &self,
+        project_id: &str,
+        project_root: &Path,
+    ) -> Result<bool, MoonError> {
+        Ok(true)
+    }
+
+    /// During project graph creation, load project aliases for the resolved
+    /// map of projects that are unique to the platform's ecosystem.
+    fn load_project_graph_aliases(
+        &mut self,
+        projects_map: &ProjectsSourcesMap,
+        aliases_map: &mut ProjectsAliasesMap,
+    ) -> Result<(), MoonError> {
+        Ok(())
+    }
+
+    /// During project creation (when being lazy loaded and instantiated in the graph),
+    /// scan for any implicit project dependency relations using the platforms manifest.
+    fn load_project_implicit_dependencies(
+        &self,
+        project_id: &str,
+        project_root: &Path,
+        project_config: &ProjectConfig,
+        aliases_map: &ProjectsAliasesMap,
+    ) -> Result<Vec<DependencyConfig>, MoonError> {
+        Ok(vec![])
+    }
+
+    /// During project creation (when being lazy loaded and instantiated in the graph),
+    /// load and infer any *additional* tasks for the platform.
+    fn load_project_tasks(
+        &self,
+        project_id: &str,
+        project_root: &Path,
+        project_config: &ProjectConfig,
+    ) -> Result<TasksConfigsMap, MoonError> {
+        Ok(BTreeMap::new())
+    }
+
+    /// Return true if the current platform is for the provided project or runtime.
+    fn matches(&self, platform: &PlatformType, runtime: Option<&Runtime>) -> bool;
+}
